feat(computeBounds): modified api of computeBounds to also use scaling - also modified & renamed tests
